chore: add form setting commands
